+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M3BR6RKNNKTF9PRX5MZ7Q9D7"),
+  id: z.string().default("01M3BY48XTB29GASBX970AJED2"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>